+++ tobas_drone/tobas_drone_msgs/include/tobas_drone_msgs/RotorConfig.hpp
@@ -16,9 +16,9 @@ struct rclcpp::TypeAdapter<tobas::RotorConfig, tobas_drone_msgs::msg::RotorConfi
   {
     dst.channel = src.channel;
     dst.link_name = src.link_name;
-    dst.direction = src.direction;
-    dst.axis = src.axis;
-    dst.esc_mode = src.esc_mode;
+    dst.direction = static_cast<uint8_t>(src.direction);
+    dst.axis = static_cast<uint8_t>(src.axis);
+    dst.esc_mode = static_cast<uint8_t>(src.esc_mode);
     dst.num_poles = src.num_poles;
     dst.max_rot_speed = src.max_rot_speed;
     dst.motor_constant = src.motor_constant;
